removing XXL as an option, since it breaks moodle
There is more work to be done here, like perhaps changing grademax to a bigger number, or making it such that the script doesn't accept XXL, but this is a start to keep people from going over 999 assignments and thus causing a failure.

I can't be sure right now how bad this failure is. Perhaps the variable can be cleared, but it does seem like it completely destroys the database. Of course, this can't be run without being put in debug mode, but nonetheless we should try to save people from this.